Refactor commands into separate files and add new commands
