+++ Animation/Assets/Scripts/Commands.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: db807b0eab5b426096be0835c51921cc
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
