+++ Animation/Assets/Scripts/Commands/ColorCommand.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f1d64e567fca4732ad9637cfdd726845
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
