+++ Animation/Assets/Scripts/Commands/JumpCommand.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7940a12eb0b34f4d8d3e78625779c8ca
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
