+++ Animation/Assets/Scripts/Commands/MoveCommand.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7b3dc284b8144a0c8a64d617f885fd55
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
