+++ Animation/Assets/Scripts/Commands/RotateCommand.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 73f71369e9024a39aaee38789b300571
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
