+++ Animation/Assets/Scripts/Commands/ScaleCommand.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3b1234edfbad4eeaadf9a1177993eebc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
