+++ Animation/Assets/Scripts/Commands/WaitCommand.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3d4d2f98fb9c41579cb3e94f84d933a2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
