+++ Animation/Assets/Scripts/RobotState.cs
@@ -4,5 +4,6 @@ public struct RobotState
 {
     public Vector3 Position;
     public Vector3 Rotation;
+    public Vector3 Scale;
     public Color Color;
 }
